@@ -14,7 +14,6 @@ classifiers =
     License :: OSI Approved :: BSD License
     Programming Language :: Python :: 3
     Programming Language :: Python :: 3 :: Only
-    Programming Language :: Python :: 3.6
     Programming Language :: Python :: 3.7
     Programming Language :: Python :: 3.8
     Programming Language :: Python :: 3.9
@@ -31,7 +30,8 @@ install_requires =
     matplotlib
     numpy>=1.13.3
     scipy
-python_requires = >=3.6
+    h5py
+python_requires = >=3.7
 
 [options.packages.find]
 exclude = test
